Refactor type handling for improved safety and consistency
- Replaced `int` with `size_t` in various loops and condition checks across multiple files, enhancing compatibility with larger datasets and ensuring safer memory management.
- Updated handling of indices in `CMenuDrawer`, `CEditView_Cmdisrch`, and `CPropTypesScreen` to utilize `size_t`, reinforcing consistent type usage.
- Documented type usage distinctions in AGENTS.md, providing clear guidelines for new code and enhancing overall type consistency across the codebase.

+++ sakura_core/apiwrap/StdControl.cpp
@@ -195,7 +195,7 @@ namespace ApiWrap{
 		std::vector<HTREEITEM> tree;
 		HTREEITEM item = TreeView_GetRoot(hwndTree);
 		while( 0 < tree.size() || item != nullptr ){
-			while(item != nullptr && (int)tree.size() < nMaxDepth ){
+			while(item != nullptr && tree.size() < static_cast<size_t>(nMaxDepth) ){
 				// 先に展開してからGetChildしないと、ファイルツリーのサブアイテムが展開されない
 				TreeView_Expand(hwndTree, item, bExpand ? TVE_EXPAND : TVE_COLLAPSE);
 				tree.push_back(item);
